Add files via upload
This file uses Q-Learning. Q-Learning helps to train a model using rewards and failures. For each right move, the model will learn the proper path, and for each failure the model will teach itself to remember the failure. 

This can be used for game development in python as opposed to pure data sceince.